mltrons/mltrons_backend/requirements.txt
amqp==2.4.0
asn1crypto==0.24.0
backports.weakref==1.0.post1
billiard==3.5.0.5
bleach==1.5.0
boto3==1.9.7
botocore==1.12.80
celery==4.2.1
certifi==2017.11.5
cffi==1.11.5
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.5
dill==0.2.8.2
Django==1.11.7
django-cors-headers==2.1.0
django-grappelli==2.12.1
django-jet==1.0.8
django-modeladmin-reorder==0.3.1
django-oauth-toolkit==1.0.0
django-rest-swagger==2.1.2
djangorestframework==3.7.3
docutils==0.14
enum34==1.1.6
funcsigs==1.0.2
future==0.16.0
futures==3.1.1
gunicorn==19.7.1
h2o==3.22.1.2
h2o-pysparkling-2.4==2.4.4
html5lib==0.9999999
idna==2.6
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
Keras==2.1.2
kombu==4.2.2.post1
Markdown==2.6.10
MarkupSafe==1.0
mock==2.0.0
ndg-httpsclient==0.5.1
np-utils==0.5.4.0
numpy==1.13.3
oauthlib==2.0.6
openapi-codec==1.3.2
pandas==0.21.0
pbr==3.1.1
Pillow==5.4.1
protobuf==3.5.0.post1
psycopg2==2.7.3.2
pusher==2.1.2
py4j==0.10.7
pyasn1==0.4.5
pycparser==2.19
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyspark==2.4.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
redis==3.0.1
requests==2.18.4
s3transfer==0.1.13
scikit-learn==0.19.1
scipy==1.0.0
simplejson==3.12.0
six==1.11.0
sklearn==0.0
tabulate==0.8.3
tensorflow==1.4.1
tensorflow-tensorboard==0.4.0rc3
Theano==1.0.1
uritemplate==3.0.0
urllib3==1.22
vine==1.2.0
Werkzeug==0.13
xgboost==0.81
